+++ packages/sakuli-legacy/src/context/common/actions/shortcut.function.ts
